4.3.0 (#85)
* Changed naming of internal components.
    * `Action` (annotated entities) is now `Activity`, not to be confused with actions that form requests. 
    * `ManualHandling` > `FunctionalHandling` - for better understanding.
* Added shortcut to `onFailure` method on `Deferred`, for consistency of approach.
* Added logging of the `Activity` list during the initialization phase.
* Internal structure of `FunctionalHandling` has been improved.
* Improved flow of coroutines in long polling mode, added request timeout in configuration, \
  see `updatesListener` block in configuration.
* Retry processing in the client has been improved, so it doesn't repeat all requests (e.g. 400 status),
  only the necessary ones.
* The type of the parameter `messageThreadId` has been changed to `Int`, as it better represents the data.
* The base scope of commands for friendliness towards beginners and better understanding has been
  changed to just `UpdateType.Message`. In other words, by default, it will have similar behavior to other types of libraries.

+++ telegram-bot/src/main/kotlin/eu/vendeli/tgbot/types/Message.kt
@@ -33,7 +33,7 @@ import java.time.Instant
 
 data class Message(
     override val messageId: Long,
-    val messageThreadId: Long? = null,
+    val messageThreadId: Int? = null,
     val from: User? = null,
     val senderChat: Chat? = null,
     override val date: Instant,
