+++ telegram-bot/src/main/kotlin/eu/vendeli/tgbot/types/internal/options/AudioOptions.kt
@@ -13,5 +13,5 @@ data class AudioOptions(
     override var disableNotification: Boolean? = null,
     override var protectContent: Boolean? = null,
     override var replyParameters: ReplyParameters? = null,
-    override var messageThreadId: Long? = null,
+    override var messageThreadId: Int? = null,
 ) : OptionsCommon, OptionsParseMode
