+++ telegram-bot/src/main/kotlin/eu/vendeli/tgbot/types/internal/options/ForwardMessageOptions.kt
@@ -3,5 +3,5 @@ package eu.vendeli.tgbot.types.internal.options
 data class ForwardMessageOptions(
     var disableNotification: Boolean? = null,
     var protectContent: Boolean? = null,
-    var messageThreadId: Long? = null,
+    var messageThreadId: Int? = null,
 ) : Options
